src/yfinance_to_kafka/requirements.txt
kafka-python>=2.0.2
yfinance>=0.2.22